+++ src/services/DatabaseTranslationsService.php
@@ -12,12 +12,17 @@ namespace digitalpulsebe\database_translations\services;
 
 use Craft;
 use craft\base\Component;
+use craft\base\Event;
 use craft\helpers\FileHelper;
+use digitalpulsebe\database_translations\DatabaseTranslations;
 use digitalpulsebe\database_translations\models\SourceMessage;
 use Exception;
+use putyourlightson\blitz\Blitz;
 
 class DatabaseTranslationsService extends Component
 {
+    const EVENT_AFTER_UPDATE = 'afterUpdateTranslations';
+
     public function languageIds(): array
     {
         return \Craft::$app->i18n->getSiteLocaleIds();
@@ -50,5 +55,32 @@ class DatabaseTranslationsService extends Component
         FileHelper::writeToFile(Craft::getAlias($file), $content);
 
         return $name;
+    }
+
+    /**
+     * Things to do after updating translations
+     * @return void
+     */
+    public function afterUpdate(): void
+    {
+        $this->trigger(self::EVENT_AFTER_UPDATE, new Event());
+        $this->clearCaches();
+    }
+    /**
+     * Clear caches for according to settings configured
+     * @return void
+     */
+    public function clearCaches(): void
+    {
+        if (DatabaseTranslations::getInstance()->settings->clearDataCache) {
+            Craft::$app->getCache()->flush();
+        }
+
+        if (
+            DatabaseTranslations::getInstance()->settings->refreshBlitzCache
+            && class_exists(Blitz::class)
+        ) {
+            Blitz::getInstance()->refreshCache->refreshAll();
+        }
     }
 }