+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 - permission for migration export format
+- Settings to clear caches after saving translations
 
 ## 2.6.5 - 2024-04-12